♻️ 작가 소개페이지 리팩토링
