+++ src/components/Photographers/Introduction/MockData/CarousalData.js
@@ -1,0 +1,5 @@
+import carousel from "../../../../assets/photograph/carousel.png";
+import carousel2 from "../../../../assets/photograph/carousel2.png";
+import carousel3 from "../../../../assets/photograph/carousel3.png";
+
+export const carousal_list = [carousel3, carousel, carousel2]; // 캐러셀 이미지
