+++ src/pages/Photographer/Review.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import ReviewBox from "../../components/Photographers/Review/ReviewBox";
 import Paging from "../../components/Photographers/Review/Paging/Paging";
-import { ReviewData } from "./MockData/ReviewData";
+import { ReviewData } from "../../components/Photographers/Review/MockData/ReviewData";
 import { useEffect, useState } from "react";
 import LayOut from "../../components/common/LayOut";
